Moved brand list query to the new directory
