+++ LanosCertifiedStore.ApplicationLayer.Application/Queries/Brands/ListQueryRelated/ListBrandsQuery.cs
@@ -3,7 +3,7 @@ using Application.RequestParams;
 using Domain.Shared;
 using MediatR;
 
-namespace Application.Queries.Brands;
+namespace Application.Queries.Brands.ListQueryRelated;
 
 public sealed record ListBrandsQuery(VehicleBrandFilteringRequestParameters RequestParameters)
     : IRequest<Result<IReadOnlyList<BrandDto>>>;
